minor refactor, added custom React hook
