+++ blog/utils/queries.ts
@@ -1,3 +1,5 @@
+import { BlogPost } from '..'
+
 export const postsQuery = `{
   alert: allAlert {
     message
@@ -71,6 +73,60 @@ export const morePostsQuery = (index: number) => `{
       title
       subtitle
     }
+    excerpt: excerptRaw
+    featured
+    mainImage {
+      alt
+      caption
+      asset {
+        metadata{
+          dimensions {
+            aspectRatio
+            width
+            height
+          }
+          lqip
+        }
+        url
+      }
+    }
+  }
+}`
+
+export const postQuery = (slug: string) => `{
+  alert: allAlert {
+    message
+    alertLink
+    internal
+    active
+  }
+  menuItems: allPage(where: {menuItem: {eq: true }}){
+    title
+    slug {
+      current
+    }
+  }
+  posts: allPost(where: {slug: {current: {eq: "${slug}"}}}) {
+    _id
+    title
+    body: bodyRaw
+    socialCard {
+      title
+      subtitle
+    }
+    slug {
+      current
+    }
+    categories {
+      title
+      slug {
+        current
+      }
+    }
+    tags {
+      value
+    }
+    publishedAt
     excerpt: excerptRaw
     featured
     mainImage {
